Refactor : 모달 알림창 분리
