+++ src/components/Comment/Comment.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import deleteComment from 'api/Comment/deleteComment';
 import reportComment from 'api/Comment/reportComment';
 import iconMoreImg from 'assets/img/icon-more-18.png';
-import CommentModal from '../common/Modal/CommentModel/CommentModal';
+import CommentModal from '../common/Modal/CommentModal';
 import ProfileImg from '../common/ProfileImg/ProfileImg';
 import {
   CommentContainerStyle,
