+++ src/components/PostCard/PostCard.jsx
@@ -2,8 +2,8 @@ import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import { accountAtom } from '_state/auth';
-import PostModal from 'components/common/Modal/PostModalAlert/PostModal';
-import PostReportModal from 'components/common/Modal/PostModalAlert/PostReportModal';
+import PostModal from 'components/common/Modal/PostModal';
+import PostReportModal from 'components/common/Modal/PostReportModal';
 import HeartIcon from '../HeartIcon/HeartIcon';
 import CommentIcon from '../CommentsIcon/CommentIcon';
 import ProfileImgAccount from '../common/ProfileImgAccount/ProfileImgAccount';
