+++ src/components/Product/Product.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import { accountAtom } from '_state/auth';
-import ProductModal from 'components/common/Modal/ProductModalAlert/ProductModal';
+import ProductModal from 'components/common/Modal/ProductModal';
 import {
   ProductWrap,
   ProductImgWrap,
