+++ src/components/common/Header/Header.jsx
@@ -4,7 +4,7 @@ import iconArrowLeft from 'assets/img/icon-arrow-left.png';
 import iconMore from 'assets/img/icon-more-18.png';
 import iconSearch from 'assets/img/icon-search.png';
 import Button from '../Button/Button';
-import LoginModal from '../Modal/LoginModalAlert/LoginModal';
+import LoginModal from '../Modal/LoginModal';
 import {
   HeaderLeftBtnStyle,
   HeaderRightBtnStyle,
